+++ src/main/java/racing/Main.java
@@ -3,7 +3,7 @@ package racing;
 import racing.domain.Racing;
 import racing.view.InputView;
 import racing.view.ResultView;
-import racing.view.Winner;
+import racing.domain.Winner;
 
 import java.util.Random;
 import java.util.Scanner;
