+++ src/main/java/racing/domain/Racing.java
@@ -1,6 +1,5 @@
 package racing.domain;
 
-import racing.view.Winner;
 import utility.NumberHelper;
 
 import java.util.*;
