+++ src/test/java/racing/RacingTest.java
@@ -2,7 +2,7 @@ package racing;
 
 import racing.domain.Car;
 import racing.domain.Racing;
-import racing.view.Winner;
+import racing.domain.Winner;
 import utility.NumberHelper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
